(double) has been missed

+++ sеm2/hw5/task1/test/hw5/task1/CalculatorTest.java
@@ -49,7 +49,7 @@ public class CalculatorTest {
     @Test
     public void testDividing() {
         double result = calculator.count(firstValue, secondValue, '/');
-        double expected = firstValue / secondValue;
+        double expected = (double)firstValue / secondValue;
         Assert.assertTrue(result == expected);
     }
 }
